move NormalizationType to own file
