+++ crossformer/data/utils/data_utils.py
@@ -35,13 +35,6 @@ def tree_merge(*trees: dict) -> dict:
     return merged
 
 
-class NormalizationType(str, Enum):
-    """Defines supported normalization schemes for action and proprio."""
-
-    NORMAL = "normal"  # normalize to mean 0, std 1
-    BOUNDS = "bounds"  # normalize to [-1, 1]
-
-
 def to_padding(tensor: tf.Tensor) -> tf.Tensor:
     if tf.debugging.is_numeric_tensor(tensor):
         return tf.zeros_like(tensor)
