+++ crossformer/data/utils/normalization_type.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class NormalizationType(str, Enum):
+    """Defines supported normalization schemes for action and proprio."""
+
+    NORMAL = "normal"  # normalize to mean 0, std 1
+    BOUNDS = "bounds"  # normalize to [-1, 1]
